Add files via upload
V1.1 adds the option to continue playing after the game tied or somebody won (one can also exit the game of course). Once the game is over, the user(s) can press Enter in order to continue playing or type quit,exit,leave or the short forms of those words (q,e,l) in order to exit the game. V1.1 also includes the fix of a potential undiscovered small bug in V1.0